Relax `Loose` associated type bound
This allows us to put a type with a smaller size into a field with
a larger size. The fit is still checked by our compile-time assertions
after monomorphization, so this should be safe. Couldn't find any
documentation about the legality of casting between integers via
union, though, which we need to keep our methods `const`.

Also looked into const trait functions, and even function pointers
as associated consts of `Loose`, but neither work--the former is
unstable and the latter doesn't allow actually calling the function
pointers in `const` contexts.